Updated Windows OS path investigation
Updated the document related to the investigation of certain paths in the Windows operating system. Added two new paths under the "Start Menu" section, which are related to the location of the Start Menu programs in the Windows file system. Also, a new "Miscellaneous" section has been added to discuss the 'Applications' folder in Windows Explorer and the need to figure out how to modify the files in this folder. This section also includes a hypothesis about the storage location of the 'Start Menu' when 'All Apps' is clicked.